third_party/blink/web_tests/external/wpt/css/css-backgrounds/background-size-001-expected.txt

This is a testharness.js-based test.
PASS background-size_initial
PASS background-size_auto
PASS background-size_cover
PASS background-size_contain
FAIL background-size_length_zero assert_equals: background-size supporting value expected "0px auto" but got "0px"
FAIL background-size_length_negative_zero assert_equals: background-size supporting value expected "0px auto" but got "0px"
FAIL background-size_length_positive_zero assert_equals: background-size supporting value expected "0px auto" but got "0px"
FAIL background-size_length_normal assert_equals: background-size supporting value expected "15px auto" but got "15px"
FAIL background-size_percentage_min assert_equals: background-size supporting value expected "0% auto" but got "0%"
FAIL background-size_percentage_normal assert_equals: background-size supporting value expected "50% auto" but got "50%"
FAIL background-size_percentage_max assert_equals: background-size supporting value expected "100% auto" but got "100%"
PASS background-size_auto_auto
PASS background-size_auto_length
PASS background-size_auto_percentage
FAIL background-size_length_auto assert_equals: background-size supporting value expected "15px auto" but got "15px"
PASS background-size_length_length
PASS background-size_length_percentage
FAIL background-size_percentage_auto assert_equals: background-size supporting value expected "50% auto" but got "50%"
PASS background-size_percentage_length
PASS background-size_percentage_percentage
Harness: the test ran to completion.
